requirements.txt
streamlit
snowflake-snowpark-python[pandas]
pandas
pypdf2